Checks board connection commands (#32)
In order to avoid errors (sw crash) when writing to serialport after
the board/serialport has been disconnected/closed additional checks are
introduced:

i) A flag (sprite.arduino.disconnecting) that marks when a board
disconnection has been initiated
ii) A function(sprite.arduino.isBoardReady()) that checks if there is a
valid board which is not been disconnected.
iii)  All commands that communicate to the board/serialport (in
s4aThreads.js) make an initial check if the board is ready.  If not an
“Arduino not connected" error is thrown.