Clears connection timeout when connected #32
A connection timeout exists to detect cases when the new Board callback
is never called (which is the case when firmata is not spoken).

If - after being connected - the board was disconnected before the
timeout had expired an error conditioned occurred.

This fix clears the timeout right after a Board is connected to avoid
this problem.